requirements.txt
torch==2.3.0
torchvision==0.18.0
pandas==2.2.0
numpy==1.26.4
tqdm==4.66.1
Pillow==10.2.0
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1
streamlit==1.36.0
gradio==4.28.0